requirements.txt
Django==4.2.11
httplib2==0.20.4
pillow==10.2.0
urllib3==2.0.7
uvicorn==0.27.1
djangorestframework
django-filter
markdown     
django-allauth[socialaccount]
django-allauth
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
django-environ